Bio::SQL::Sequence#to_biosequence is changed to use "adapter", but not tested
New module Bio::Sequence::Adapter::BioSQL for Bio::SQL::Sequence to
Bio::Sequence adapter module, and Bio::SQL::Sequence#to_biosequence
is changed to use the adapter.

Note that the changes in this commit is NOT tested whether to work
correctly, because I don't have BioSQL environment.

(Modified in 2016/6/17)
Changes to the below files were included in the original commit
437132f6582eda4411a605866308c3d9b5cf4285
but were not included in this commit:
	lib/bio/sequence/adapter.rb